Moving attendance from views to components
